Integrate FE with BE APIs - Except for Photo

+++ src/main/java/ng/first/app/be/service/StudentService.java
@@ -68,7 +68,9 @@ public class StudentService {
 		
 		boolean isDeleted = false;
 		try {
-			studentRepository.deleteById(id);
+			if (studentRepository.findById(id).isPresent()) {
+				studentRepository.deleteById(id);
+			}
 			isDeleted = true;
 		} catch(Exception ex) {
 			ex.printStackTrace();
